Fix search feature in assignment2
