+++ assignment-2/public/index.html
@@ -28,6 +28,10 @@
       rel="stylesheet"
       href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap"
     />
+    <script
+      src="https://kit.fontawesome.com/372c90f6b7.js"
+      crossorigin="anonymous"
+    ></script>
     <title>Bookstore</title>
   </head>
   <body>
